src/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
colorama==0.4.6
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
pyrogram==2.0.106
pyTelegramBotAPI==4.22.1
requests==2.32.3
pytest==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1
aiogram==3.27.0
asyncpg==0.31.0